API code connecting to Open Notify
There are other API endpoints here for twitter, github and even gmail. They can all be used as is with your own personalised oauth code. Without oauth you will get 400/404 responses.
The Open notify code can be used as is. It pulls a list of all astronauts in space and orders them in a list and then outputs both.